+++ tests/test_providers.py
@@ -33,7 +33,7 @@ larch = {
     ],
     'member_of': ['3'],
     'matches': {
-        'close': 'http://id.python.org/different/types/of/trees/nr/1/the/larch'
+        'close': ['http://id.python.org/different/types/of/trees/nr/1/the/larch']
     }
 }
 
@@ -51,7 +51,10 @@ chestnut = {
             'note': 'A different type of tree.'
         }
     ],
-    'member_of': ['3']
+    'member_of': ['3'],
+    'matches': {
+        'related': ['http://id.python.org/different/types/of/trees/nr/17/the/other/chestnut']
+    }
 }
 
 species = {
@@ -226,7 +229,10 @@ class TreesDictionaryProviderTests(unittest.TestCase):
         self.assertEqual(larch['notes'], lariks.notes)
         self.assertEqual(larch['member_of'], lariks.member_of)
         self.assertEqual('concept', lariks.type)
+        assert 5 == len(lariks.matches)
+        assert 1 == len(lariks.matches['close'])
         assert larch['matches']['close'] == lariks.matches['close']
+        assert [] == lariks.matches['related']
 
     def test_get_by_id_is_type_agnostic(self):
         self.assertEqual(trees.get_by_id(1), trees.get_by_id('1'))